[#14] Reduce the overall number of jobs to run on Travis CI.
Most Python versions only test against the most recent pytest and
Cython.

The most recent Python version is tested against a wider range of
configurations.  This still has a small probability of something
slipping through the cracks, but we can worry about that if and when
something is actually discovered.