New patch for py.path.pyimport / import_path which does not require copying as much code from pytest.
This started as simplifications that can be made on the basis of
dropping Python 2 support (#15) namely that we can pretty well
rely on the sysconfig variable EXT_SUFFIX (though we could also use
importlib) to get the ABI-tagged extension module suffix.  This works
just as well on PyPy 3.x too.

I later realized that we can hackily work around pytest's lack of
support for this by wrapping the path objects (py.path.local for
earlier versions of pytest that still use py.path, or pathlib.Path for
newer versions) in such a way that they can ignore the ABI tag part of
the file extension while correctly determining the module name from the
filename.  This is a bit fragile, but requires less code
re-implementation (e.g. support for import_mode='importlib').

Also the code in pytest that this works with (py.path.local.pyimport and
import_path) are not likely to change significantly, especially on the
older versions).  For newer versions of pytest I will try to get better
support for extension modules merged upstream.