More fixes to ensure that if a docstring is missing from a method, but the doctest is still available in the __test__ dict, use the copy of the test from the __test__ dict transparently.
This is especially needed on PyPy since there is a limitation of
Cython on PyPy that docstrings on special methods are not accessible
from the Python level (so in the tests Eggs.__init__ was not being
tested).

Now it takes the test from the __test__ dict but modifies the name
of the test (and sets the line number) to appear as if the docstring
is being tested directly from the method.